Use PoW BlockHeader hash as next block previous_hash
We want to use the hash resulting from PoW as previous_hash.
This hash now completely constrain the BlockHeader, and the BlockHeader completely constrain the `transactions`, so we can still verify the block data is correct only with previous_hash even if we now
need to deserialize the BlockHeader.

- Fix api tests so the block headers have correct values (hash and nonce).
- Use construct_valid_block_pow_hash everywhere for consistency to check block is validatable and get the hash.

Test:
Add validation that SotredSerializedBlock can be validated
cargo clippy
cargo test