Parralelize participants and block PoW
The goal is to take a big step toward the target design for parallel
PoW.
Parallelizing at a higer level is very problematic in regards to
reproductability, testing and clarity.

In order to simplfy the flow, both PoW are triggered by the same
messages.
In order to handle restarting the network after upgrade or on first
start we also have a first step gathering
participants first if they are not already present.

Handle late partition Pow due to double pow for first
Adjust test to avoid random failure without changing the response for
PoW that come too late

Test:
cargo clippy
cargo test